Added cleanData() and modified Spell_demo
In Spell_Demo you can specify a list of regex expressions (just like you do with the preprocessing). This list will be used to pre clean the data before the Spell algorithm starts. For example, pytest logs contain strings such as "-----" or multiple ":" characters. These seem to confuse Spell and it treats some of them as parameters therefore leading to real parameters not being parsed. 

I think this might be useful if developers know already some "noise" patterns that appear in the logs and want to remove them